refactor(utils): move the utils into the right place
