+++ src/dashboard/apigateway/apigateway/apps/monitor/tasks.py
@@ -23,6 +23,7 @@ from typing import Any, Dict
 
 import arrow
 from celery import shared_task
+from django.conf import settings
 
 from apigateway.apps.monitor.constants import (
     ALARM_RECORD_RETENTION_DAYS,
@@ -32,7 +33,6 @@ from apigateway.apps.monitor.constants import (
     NoticeWayEnum,
 )
 from apigateway.apps.monitor.models import AlarmRecord
-from apigateway.apps.monitor.utils import get_es_index
 from apigateway.service.alert_flow.handlers.app_request import AppRequestAlerter, AppRequestAppCodeRequiredFilter
 from apigateway.service.alert_flow.handlers.base import (
     AlarmRecordCreator,
@@ -47,6 +47,11 @@ from apigateway.service.alert_flow.handlers.resource_backend import (
 from apigateway.service.alert_flow.helpers import AlertFlow, MonitorEvent
 
 logger = logging.getLogger(__name__)
+
+
+def get_es_index(alarm_type: AlarmTypeEnum) -> str:
+    config = getattr(settings, "BKMONITOR_ALARM_CONFIG", {})
+    return config[alarm_type.value]["es_index"]
 
 
 @shared_task(name="apigateway.apps.monitor.tasks.monitor_resource_backend")
