+++ src/dashboard/apigateway/apigateway/apps/permission/managers.py
@@ -27,7 +27,7 @@ from apigateway.apps.permission.constants import (
     GrantDimensionEnum,
     GrantTypeEnum,
 )
-from apigateway.apps.permission.utils import calculate_expires, calculate_renew_time
+from apigateway.common.time import calculate_expires, calculate_renew_time
 from apigateway.utils.time import now_datetime, to_datetime_from_now
