+++ src/dashboard/apigateway/apigateway/editions/ee/apps/esb/management/commands/sync_esb_permissions_to_gateway.py
@@ -31,7 +31,7 @@ from apigateway.apps.esb.exceptions import EsbGatewayNotFound
 from apigateway.apps.esb.utils import get_esb_gateway
 from apigateway.apps.permission.constants import GrantTypeEnum
 from apigateway.apps.permission.models import AppResourcePermission
-from apigateway.apps.permission.utils import calculate_expires
+from apigateway.common.time import calculate_expires
 from apigateway.core.models import Gateway
 
 logger = logging.getLogger(__name__)
