+++ src/dashboard/apigateway/apigateway/service/utils.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -15,3 +16,27 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
+
+from django.conf import settings
+
+
+def get_resource_url(resource_url_tmpl: str, gateway_name: str, stage_name: str, resource_path: str):
+    """
+    拼接资源请求地址
+    """
+    return resource_url_tmpl.format(
+        api_name=gateway_name,
+        stage_name=stage_name,
+        resource_path=resource_path.lstrip("/"),
+    )
+
+
+def get_resource_doc_link(api_name, stage_name, resource_name):
+    """
+    资源文档链接
+    """
+    return settings.RESOURCE_DOC_URL_TMPL.format(
+        api_name=api_name,
+        stage_name=stage_name,
+        resource_name=resource_name,
+    )
