--- src/dashboard/apigateway/apigateway/tests/apps/mcp_server/test_utils.py
@@ -1,28 +1,0 @@
-#
-# TencentBlueKing is pleased to support the open source community by making
-# 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
-# Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-# either express or implied. See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# We undertake not to change the open source license (MIT license) applicable
-# to the current version of the project delivered to anyone in the future.
-#
-
-
-from apigateway.apps.mcp_server.utils import build_mcp_server_url
-
-
-class TestBuildMCPServerURL:
-    def test_build_mcp_server_url(self, settings):
-        settings.BK_API_URL_TMPL = "http://test.com/{api_name}"
-
-        url = build_mcp_server_url("test-mcp-server")
-        assert url == "http://test.com/bk-apigateway/prod/api/v2/mcp-servers/test-mcp-server/sse/"
